refactor: 검색 인풋 위치 수정 (FE-134)
