+++ src/components/board/ArticleSection/index.tsx
@@ -1,40 +1,21 @@
 import { Article } from "@coworkers-types";
 import Dropdown from "@components/commons/Dropdown";
-import Input from "@components/commons/Input";
 import Label from "@components/commons/Label";
 import ArticleCard from "./ArticleCard";
 
 export default function ArticleSection({
   Posts,
-  searchValue,
-  searchChange,
-  onEnter,
   sortValue,
   sortChange,
 }: {
   Posts: Article[];
-  searchValue: string;
-  searchChange: (value: string) => void;
-  onEnter: (event: React.KeyboardEvent<HTMLInputElement>) => void;
   sortValue: string;
   sortChange: (value: string) => void;
 }) {
-  const handleKeywordChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    searchChange(event.target.value);
-  };
   return (
     <div className="mt-30">
       <div className="mb-12 flex w-full items-center justify-between gap-30">
         <Label content="게시글" className="w-100" />
-        <Input
-          value={searchValue}
-          onChange={handleKeywordChange}
-          onKeyDown={onEnter}
-          type="search"
-          name="search"
-          placeholder="검색할 게시글을 입력해주세요."
-          className="flex w-600 items-center justify-center"
-        />
         <Dropdown selectedValue={sortValue} onSelect={(value) => sortChange(value)}>
           <Dropdown.Toggle>최신순</Dropdown.Toggle>
           <Dropdown.Wrapper>
