+++ src/components/board/BestArticleSection/index.tsx
@@ -6,7 +6,7 @@ import BestArticleCard from "./BestArticlesCard";
 
 export default function BestArticleSection({ Posts }: { Posts: Article[] }) {
   return (
-    <div className="mx-0 my-auto flex w-full flex-col gap-12">
+    <div className="mx-0 my-auto mt-30 flex w-full flex-col gap-12">
       <div className="flex items-center justify-between">
         <Label content="베스트 게시글" />
         <Link href="/add-board">
